OC10-042 fix incorrect dynamic invariant
There was a mismatch between the condition computed before calling
New_Is_Constrained_Access and the condition computed inside
New_Is_Constrained_Access. This mismatch could result in some cases in a
dynamic invariant of the form True = False.

* gnat2why-expr.adb
(Compute_Dynamic_Invariant): fix computation of type to be passed to
  New_Is_Constrained_Access