OC15-027 flow - fix crash
When processing "foo.bar", we need to be aware that this can be a call
into a protected object. We already dealt with entry and function calls
but for some reason we were missing the case where we can call a
procedure. This would not happen during normal operation, but when
emitting an error message we call Get_Variable_Set on all sorts of odd
constructs in order to determine a good node for the message...