OC16-015 stats for packages with external axioms
Statistics (spec/body in/out SPARK) are now reported for packages with
annotated with External_Axiomatization.

* gnat2why/spark/spark_definition.adb
(Mark_Package_Declaration, Mark_Package_Body): generate statistics for
  packages with external axiomatization.

* testsuite/gnatprove/tests/OC16-015__ext_axioms_stats/
Test enabled and output added. We check for "all" status since package there
is no package body file and effectively 'all' of package is in SPARK. Also,
there is no need for a client of the package.