P111-029 do not delay marking of type predicates
Type predicates are now marked just after their corresponding types and
within the context of the same SPARK_Mode pragma. This avoids a "delay" in
marking of type that required troublesome storing the type's SPARK_Mode
pragma.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): do not delay marking of type predicate until the end of
  compilation unit.

(Mark_Entity): mark type predicate just after marking the type itself and
  within the context of the same SPARK_Mode pragma.

* testsuite/gnatprove/tests/O416-002__dynamic_predicate_not_in_spark/
Test output updated. Now we do not complain if a type declared in package
with SPARK_Mode => Auto is not in SPARK because of SPARK violations in its
predicate expression.

* testsuite/gnatprove/tests/P111-029__marking_delayed_aspect/
Tests that specifically check for regressions in this TN.