P112-047 marking of record components
When marking record components point to location where a non-SPARK type is
used, not to such a type itself. Now the user can easily navigate to the
problematic type; previously he was pointed to type declaration and had to
guess from where the type has been referenced.

Also, if the non-SPARK type is used in several records we a message for each
reference. Previously, we were giving only one message and the user had to
iteratively remove references to the non-SPARK type.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): attach violations in types of record components to
  fields, not to types.

* testsuite/gnatprove/tests/P112-047__record_not_in_spark/test.out
Test output updated.