P111-029 show crash on new marking of predicates
This commit fixes the new code for marking of delayed predicates so that the
tool wrongly classifies a type as in SPARK and then crashes in proof.

This shows why the code must be reverted.

* gnat2why/spark/spark_definition.adb
(Mark_Subprogram_Body): missing piece in the new (and wrong) marking;
  without marking of the predicate expression function body the new scheme
  did not really looked for violations in the predicate expression.

* testsuite/gnatprove/tests/P111-029__delayed_aspects/
Test extended to show the problem and disabled to avoid the crash.