P114-035 post SPARK violations on formals
Marking violations should be posted on the immediate entity that is not in
SPARK, not on the entity which will be not in SPARK due to violation.

For example, if a function has several formal parameters and only one of
them is not in SPARK we should point exactly to the offending one and not to
the function itself. Otherwise it is impossible to figure out where is the
violation.

Interestingly, this makes marking of formals in subprogram specs just like
in bodies and there is definitely no reason to mark them differently.

Note that this patch only removed error messages in the testsuite, but this
is temporary (hopefully). Now if the type of a formal parameter is not in
SPARK and we already posted an error about the type itself then GNAT will
prevent any errors posted on entities of that type. This most likely
deserves a separate TN.

* gnat2why/spark/spark_definition.adb
(Mark_Subprogram_Specification): post SPARK violations in formals right on
  formals themself.

* testsuite/gnatprove/tests/*/test.out
Test outputs updated.

* testsuite/gnatprove/tests/P114-035__error_on_func_return_type/p.ads
Text extended.