P115-045 test for predicate on type instance
Front end already processes this code correctly, but now we crash in
GNATprove. It is because when we see Has_Predicates flag we expect
Predicate_Function to return non-empty value. This test is similar to
OC12-001, but here we should emit a check.

* testsuite/gnatprove/tests/P115-045__predicate_on_type_instance/
Test similar to OC12-001, but here we should emit a check. Currently
crashes.