P112-038 marking of private type with predicate
Fix marking of private types with predicates by correctly marking
declaration of function that is generated by the front end and implements
the predicate expression.

* gnat2why/spark/spark_definition.adb
(SPARK_Pragma_Of_Type): lifted to library-level subroutine.

(Mark_Call): calls to predicate functions that are not in SPARK should be
  reported as references to their corresponding types.

(Mark_Subprogram_Declaration): detect when marking declaration of function
  that implements a type predicate.

(Mark_Subprogram_Body): if body of a function generated by the front end to
  check the predicate expression is not in SPARK then remove the
  corresponding type from the set of entities in SPARK.

* testsuite/gnatprove/tests/P112-038__predicate_private_type_not_in_spark/
Extend test with a client package that makes sure that the code has been
marked as in SPARK.