OC13-004 Do not generate spurious predicate check on predicate expr
A spurious predicate check was generated on internal calls to the
predicate function applying to a parent type of the current type, when
checking the absence of run-time errors in a predicate expression.
This special case is now recognized, and no check is issued on the type
conversion from the derived type or subtype to the parent type in that
case.

* gnat2why/spark/spark_util.adb|s
(Is_Call_Arg_To_Predicate_Function): Recognize internal call to parent
  predicate function.
* gnat2why/why/why-gen-expr.adb
(Insert_Array_Conversion,
 Insert_Record_Conversion,
 Insert_Scalar_Conversion): Do not insert predicate check in the special
  case of an explicit call to the predicate function, corresponding to
  the definition of a the predicate function of a subtype or derived type.
* testsuite/gnatprove/tests/OC13-004__predicate/test.out
Fix test output.