minor remove duplicated check for volatile formals
Volatile formals are already detected in by the front end (where SPARK_Mode
is On) and marking (where SPARK_Mode is Auto). There is no need to duplicate
this check in flow, especially with a message that says about "global" and
not "formal."

Also, now the code implements a specific SPARK RM rule and nothing else.

* gnat2why/flow/flow-analysis.adb
(Check_Function_For_Volatile_Effects): do not duplicate checks formals, this
  was already done by the front end and marking.