P202-010 fix problem of running SPARK on windows in some cases
In some cases SPARK analysis on Windows will stop with message

gnat2why: RTS path not valid: missing adainclude and adalib directories

when GNAT is installed in the same location. This is due to a bug in
gnatcoll API (see P211-001) when querying Runtime and Target Attributes,
but in fact there is a better/simpler way of querying these attributes,
which also avoids the bug.

* configuration.adb
(Set_RTS_Dir): use Get_Target and Get_Runtime instead of querying the
  attributes manually