P226-022 simplify querying SPARK_Mode of an Itype
SPARK_Mode of an Itype is determined by the SPARK_Mode that applies to where
its associated type is, not of the SPARK_Mode that applies to where the
Itype itself is put by the compiler.

* gnat2why/spark/spark_definition.adb
(Mark_Entity, Is_Private_Entity_Mode_Off): retrieve SPARK_Mode of Itype's
 associated type even for itypes that have their grandparent node set.

(Is_Private_Entity_Mode_Off): precondition added to document the interface.