P229-022 no flow warning on exceptions
Flow analysis does not longer warn about unused initialization for variables
whose values are used to decide whether to raise an exception.

An exception is an externally visible outcome, just like assignment to an
output variable; thus data that decide whether to raise an exception is
"used" as much as those that decide what are the values of output variables.

* gnat2why/flow/flow-analysis.ads
* gnat2why/flow/flow-analysis.adb
(Is_Final_Use_Any_Export): detect also normal use that may branch to a raise
 statement.

* testsuite/gnatprove/tests/P229-022__unused_with_exceptions/
Test enabled and extended; output added.