P229-022 no warn on imports that affect exceptions
Suppress "unused initial value" warning on global variables that are inputs
to subprograms (aka. "imports") and are used to (potentially) branch to
exceptional paths (e.g. paths that lead to raise statements, statically
false assertions or calls to subprograms with pragma No_Return).

This is consistent with the suppressing a similar warning on local variables
that are used for the same purpose.

* gnat2why/flow/flow-analysis.adb
(Is_Final_Use): detect also vertices that correspond to points of branching
 to exceptional paths.

* testsuite/gnatprove/tests/P229-022__unused_with_exceptions/test.out
Main test output for the TN updated.

* testsuite/gnatprove/tests/N701-018__flow_raise_statements/test.out
Output for other test also updated.