P218-017 explicitly disallow cache on some tests
Some tests inherently rely on the output generated by gnatwhy3 and always
result in diffs when run with the --cache option, so should never be run
with it.

* testsuite/gnatprove/lib/python/test_support.py
(prove_all): new option "cache_allowed" which defaults to True, but can be
 explicitly set to False in driver files for tests that should never use
 caching mechanism.

* testsuite/gnatprove/tests/*/test.py
Disable cache on tests that are inherently incompatible with caching.