minor simplification of graph Parent operation
It feels more natural to get the first element of a singleton set by calling
First rather than looping over the set just to finish at the very first
iteration.

This change does not break any of the abstractions provided by the package,
because the simplified procedure belongs to the package body and thus has
full view of the internal graph representation. Also, other subprograms in
the same package body loop directly over the In_Neighbours set, not over
the abstraction provided by a collection with the same name.

* gnat2why/flow/graphs.adb
(Parent): simplify retrieval of the first element of a set.