renamings instead of expensive local copies
Vertex attributes represented by type V_Attributes are composed of several
sets and so it may be expensive to create local copies of this type. Rather
than copies, it seems better to use renamings, which are optimized as
references by the compiler (and whose declarations are less verbose for the
human).

Also, renamings are preferred when the intention is just to mutate an
element of a container, i.e. instead of:

   declare
      Local_Copy : Element := Some_Set (Some_Index);
   begin
      Local_Copy.Some_Component := New_Value;
      Some_Set (Some_Index) := Local_Copy;
   end;

it seems better to write:

   declare
      Some_Element : Element renames Some_Set (Some_Index);
   begin
      Some_Element.Some_Component := New_Value;
   end;

which is easier to understand and cheaper to execute (it saves us both from
creating a local copy and doing extra search in the container when inserting
the updated element).

(no-tn-check)