cleanup of calls to Element on containers
Calls to Element make code more verbose and result in a lot of unnecessary
local copies. Renamings with implicit references are easier to read and more
efficient to execute.

(no-tn-check)

Note of justification: so far I have not seen any call to Element where the
programmer's intent was to create a local copy; in all cases the intent was
to create a shorter name to access the contents of a container. This is
pricesely what renamings and references are for.

Also, it seems right to cleanup this before starting a much heavier
re-implementation of global generation.