P301-005 Add checking of predicate on default initialization value
No predicate check was performed by GNATprove on the default
initialization of an object (of a type with full default initialization)
subject to a predicate or whose subcomponents were subject to
a predicate. This is now performed as part of the checking of the default
initialization value assigned to such an object. Note that the check is
located on the type, but it is only triggered if an object of the type
is default initialized. This is similar to the checks for default values.

This patch makes it useless to keep the previously introduced limitation
that null record types with predicates are not supported in SPARK. Hence
we also lift this limitation at the same time.

* common/vc_kinds.ads
(VC_Kind): Add new VC kind for predicate check on default value.

* docs/ug/gnatprove.rst
Update doc of VC kinds.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): Remove limitation on null record types with
  predicates.

* gnat2why/spark/spark_util.adb|s
(Find_Predicate_Aspect): New function to return applicable predicate
  if any.

* gnat2why/why/gnat2why-error_messages.adb
* gnatprove/spark_report.adb
Adapt to new VC kind.

* gnat2why/why/gnat2why-expr.adb
(New_Predicate_Check): Add parameter to distinguish between predicate
  check on value, and predicate check on default value of type.
(Compute_Default_Check): Add check on predicate to the checking of default
  initialization value.

* testsuite/gnatprove/tests/P115-045__predicate_on_type_instance/test.out
* testsuite/gnatprove/tests/P118-002__predicate_component/test.out
* testsuite/gnatprove/tests/P126-024__record_predicate/test.out
* testsuite/gnatprove/tests/P301-005__pred_rec_decl/preds.adb
* testsuite/gnatprove/tests/P301-005__pred_rec_decl/preds.ads
* testsuite/gnatprove/tests/P301-005__pred_rec_decl/test.out
Update test output.

* why3
Update Why3 version for new VC kind.