P301-012 Fix missing predicate check on loop bounds
Because of the way predicate checks are inserted, they could be missing
on scalar loop bounds. Now fixed.

* gnat2why/why/gnat2why-expr-loops.adb
(Transform_Loop_Statement): Insert declaration of loop bounds in right
  order (minor adjustment).

* gnat2why/why/gnat2why-expr.adb
(Transform_Expr): Force type conversion to type with predicate to go
  through predicated type, instead of relying on Insert_Checked_Conversion
  to detect need for predicate check in such a case.

* testsuite/gnatprove/tests/P301-012__pred_type_in_loop/test.out
Test passes now.

* testsuite/gnatprove/tests/K607-022__unconstrained_array/test.out
* testsuite/gnatprove/tests/NA11-009__very_long_subunit/test.out
* testsuite/gnatprove/tests/O116-027__very_long/test.out
* testsuite/gnatprove/tests/OB04-007__float_conversion/test.out
Update output with missing checks.

* testsuite/gnatprove/tests/O416-002__dynamic_predicate/test.out
Duplicate predicate check is emitted in rare cases. This is better than
skipping predicate checks, so we keep it like that for now. The reason is
that here a predicate check is inserted on the expression being converted
by Insert_Checked_Conversion, and then the same predicate check is
inserted (due to our change) on the type conversion node itself.