fix of useless iterations in Get_Graph_Closure
When Get_Graph_Close was called with a node map like:

 X -> {X, Y, Z}

it copied the keys of the map (X) into a working set, then took first
element (X) from it, deleted this element from the working set, and finally
collected all corresponding items from its the map (X, Y, Z) thus inserting
X back into the working set.

This resulted in an extra, completely useless iteration with X. This is
fixed in this commit.

(no-tn-check)

* gnat2why/why/gnat2why-util.adb
(Get_Graph_Close): fix when the element is deleted from the working set to
 avoid use iterations.