OA13-020 Use stable names for aggregates used to define objects
Aggregates in SPARK get transformed into functions in Why3, with a unique
name based on an internal counter in GNATprove. This is not ideal for
aggregates used to define objects which are used in interactive proofs,
as the name of the aggregate may change from one version of GNATprove to
the next. Special case these aggregates so that their name is uniquely
based on the name of the object they define.

* gnat2why/why/gnat2why-expr.adb
(Transform_Aggregate): Return stable name for aggregate when used to
  define an object.

* gnat2why/why/why-atree-modules.adb|s
(Get_Module_Name): New function to retrieve the name of a module.

* gnat2why/why/why-gen-names.adb|s
(Why_Name_Enum): New identifier for aggregate definition suffix.

* testsuite/gnatprove/tests/OA13-020__aggregate_name/aggr.ads
* testsuite/gnatprove/tests/OA13-020__aggregate_name/test.out
* testsuite/gnatprove/tests/OA13-020__aggregate_name/test.py
New test. The actual name in Why3 is not tested, only that the code can be
proved.