OA26-024 Add display of counterexample traces in tests
It is now possible to display counterexample traces in a human readable
form in the expected output of tests. One simply has to add a call to
check_counterexamples() in test.py, as well as markers @COUNTEREXAMPLE in
the code on lines where a counterexample is expected and its trace should
be displayed in the output.

Currently, we use it on five representative tests from Riposte.
This feature should be used whenever we want to ensure a given
counterexample trace continues to be issued for a given proof failure.

* testsuite/gnatprove/lib/python/test_support.py
(check_counterexamples): New function to detect missing counterexamples
  and display counterexample traces.

* testsuite/gnatprove/tests/riposte__arithmetic/arithmetic.adb
* testsuite/gnatprove/tests/riposte__arithmetic/test.out
* testsuite/gnatprove/tests/riposte__arithmetic/test.py
* testsuite/gnatprove/tests/riposte__arrays/arrays.adb
* testsuite/gnatprove/tests/riposte__arrays/test.out
* testsuite/gnatprove/tests/riposte__arrays/test.py
* testsuite/gnatprove/tests/riposte__basic/basic.adb
* testsuite/gnatprove/tests/riposte__basic/test.out
* testsuite/gnatprove/tests/riposte__basic/test.py
* testsuite/gnatprove/tests/riposte__enums/enums.adb
* testsuite/gnatprove/tests/riposte__enums/test.out
* testsuite/gnatprove/tests/riposte__enums/test.py
* testsuite/gnatprove/tests/riposte__records/records.adb
* testsuite/gnatprove/tests/riposte__records/test.out
* testsuite/gnatprove/tests/riposte__records/test.py
Add counterexample display on 5 representative tests.