NB04-027 Adapt LSP checking to correct semantics of classwide pre/post
AI12-0113 has changed the semantics of primitive calls to controlling
arguments in classwide pre/post, so that these calls are not dispatching
calls anymore, but are reinterpreted at the level where the classwide
pre/post is inherited. The frontend is creating the specialized versions
of the inherited classwide pre/post. GNATprove is now using these
specialized versions instead of the ones on the overridden operation.

* gnat2why/spark/spark_util.adb
(Find_Contracts): Look for inherited classwide pre/post in Contract node
  attached to subprogram entity, instead of looking at contracts of
  overridden operations.

* gnat2why/why/gnat2why-expr.adb
* gnat2why/why/gnat2why-subprograms.adb
* gnat2why/why/why-gen-binders.ads
Fix coding style.

* testsuite/gnatprove/tests/N709-001__contracts/test.out
* testsuite/gnatprove/tests/O217-033__stacks/test.out
* testsuite/gnatprove/tests/O217-033__stacks/test.py
* testsuite/gnatprove/tests/O504-014__state_machine/test.out
Update test outputs.