P330-001 Generate axiom for expression function in SPARK_Mode => Auto
When an expression function in a part of code with no explicit SPARK_Mode
set was called in SPARK code, the axiom generated for its body was not
available for proof. This could be surprising, as the postcondition of
a function in a similar case is available for proof. The difference of
treatment between those two cases has been removed, so that an axiom is
now generated for the body of the expression function in such a case.

* gnat2why/spark/spark_definition.adb|s
(Bodies_Compatible_With_SPARK): New global variable to store the set of
  expression functions whose body can be transformed into an axiom in
  Why3.
(Entity_Body_Compatible_With_SPARK): New function to access above
  variable.
(Mark_Subprogram_Body): Mark the body of expression functions when not
  in SPARK_Mode => Off, and record compatible bodies in the new variable
  above.

* gnat2why/why/gnat2why-driver.adb
(Complete_Declaration): Use new function.

* gnat2why/why/gnat2why-subprograms.adb
(Translate_Expression_Function_Body): Use new function.

* testsuite/gnatprove/tests/P330-001__sparkmode/test.opt
* testsuite/gnatprove/tests/P330-001__sparkmode/test.out
Test passes now.