P328-010 commit minor test diffs in release branch
these are only diffs in counter examples, very minor proof result diffs.

The "compilation failed" message in P113-023__alias_issue is in fact
expected, because the change in gprbuild that removes this message is
not in the release. And the test output never contained the message
because it has been added on the devel branch after branching for the
release.