minor restriction of run-time trampolines
Ideally, gnat2why should be build with the same restrictions as is the front
end, so that code can be easily moved between them (e.g. from spark_util.adb
to sem_util.adb). (It is only realistic for some features, since gnat2why is
in Ada 2012 and front end is not).

This patch prevents one such feature that is already prohibited in the front
end. Also, it prevents the code that is "quite slow at runtime" and does not
work "on machines that prevent execution of stack data" (e.g. windows with
DEP, or data execution protection) [quotes from GNAT RM].

* gnat2why/gnat.adc
(No_Implicit_Dynamic_Code): disallow use of trampolines.