OB19-001 dynamic bounds in function contracts
Proof used to crash when the first or last attribute of
a dynamic type were used in a function's contract.
That is because the logic function for the attribute
was expecting references as arguments whereas, in the
axiom generated for function contracts, global variables
are replaced by (universally quantified) variables.

* gnat2why/why/why-gen-binders.adb
(Get_Parameters_From_Binders): Do not expect regular
binders to always be mutable.
(Get_Args_From_Variables): Only dereference binders if
they are mutable.