minor fix of CFG loop context management
When building the context related to nested loops is managed as a stack: at
the loop entry some information is added, at the loop exit this information
should be dropped (but was not). This is now fixed.

* gnat2why/flow/flow-control_flow_graph.adb
(Do_Loop_Statement): when starting a loop insert (instead of including) an
 empty set with references, because inserting gives an extra uniqueness
 check; fix useless name of local variable to something meaningful; delete
 map with loop references once loop processing is done.

(Create): assert that once the CFG is done then all loop context information
 is dropped.