minor split of code for empty/non-empty list
This code was quite subtle, because there were two local variables: one
required if the input is empty, other if it is non-empty. Also, the initial
assignment to Block was useless.

Now the empty/non-empty cases are handled separately, there are no useless
assignments and no special-casing in the loop body for a one-element
list (which was hard to follow and slow to execute).

* gnat2why/flow/flow-control_flow_graph.adb
(Join): rewritten.