OB17-003 enable tests since problem is fixed
Enable tests for OB12-012 and O909-010, since the blocking TN is now fixed.

Marking of assertions in these tests as required was wrong, because entry
body should not assume that the contract on its protected components holds
when the body is executed (in can only assume that this contract holds on
the point of call).