OC10-050 fix missing check for POs nested in other data
Previously, we were checking some properties of protected objects at
their declaration, like this:

  PO : Protected_Type;

but the code wasn't triggered when instead a record with fields of
protected type was declared, or an array.

Instead, we now check properties of protected objects directly on the
type.

This leads to some properties being unproved, because they are now
checked in a context with less knowledge about each object. Also, the
messages move to the proctected type instead of the object declaration.

* spark_util.ads
weakened some preconditions to allow for more use cases
* gnat2why-driver.adb
(Do_Generate_VCs): now also process protected types similar to task
  types
* gnat2why-expr.adb
(Transform_Declaration): remove most code for protected types
* gnat2why-subprograms.adb
(Generate_VCs_For_Protected_Type): new function to generate VCs related
to protected types. This includes check that interrupt is reserved, and
priority is in range.