P318-019 postorder traversal of entity tree
In marking the entities are inserted into a tree and in flow analysis they
should be traversed in what Knuth describes as "postorder" fashion, i.e.
children first and then the parent. This is precisely what the Rationale for
Ada 2012(8.4) describes as the "depth-first" order; however, ARM mandates
that the "depth-first" order is what Knuth calls "preorder" and this is what
GNAT provides.

This patch adds a recursive procedure Iterate_Entities that implements the
"postorder" that is needed for new generated globals.

* gnat2why/flow/flow.adb
(Build_Graphs_For_Entity): refactored as a local procedure (originally in
 Build_Graphs_For_Compilation_Unit); the "others" case-branch changed to
 raise an exception, since now we iterate only over interesting entities.

* gnat2why/spark/spark_definition.ads
* gnat2why/spark/spark_definition.adb
(Iterate_Entities): new procedure to implement the "postorder" traversal.

(Entity_Tree): moved in body, since the "preoder" iteration provided by GNAT
 cannot be used anyway.