P415-005 test - smtlib case study
Add lexer for smtlib (2.5) with absence of RTE proven. We also maintain
a small invariant, but the bulk of the interesting proof is in read_token
where we post-process some tokens (and thus need to reason about string
length of any particular token).

Termination proof would be nice in the future.