P418-016 Generate dummy value for private type
Private type is handled like a special record in Why3, except no dummy
value was generated. This lead to a crash when creating a value of a variant
type, as this could lead to a case where the dummy value was required for
the private type. Now fixed by generating a dummy value also for private
types.

* gnat2why/why/gnat2why-types.adb
(Translate_Type): Generate dummy value for private type as well.

* testsuite/gnatprove/tests/P418-016__spark_mode_off/lexer.adb
* testsuite/gnatprove/tests/P418-016__spark_mode_off/names.adb
* testsuite/gnatprove/tests/P418-016__spark_mode_off/names.ads
* testsuite/gnatprove/tests/P418-016__spark_mode_off/test.out
* testsuite/gnatprove/tests/P418-016__spark_mode_off/test.py
Add test.