P318-019 fix order of traversal for new GG
Yet another fix of the order of traversal. Initially a single multi-tree
container seemed to be a neat solution (because iteration routines are
already there); now it seems too complicated and error prone (first, because
standard iteration does preorder traversal while we need postorder and so we
have our own iteration anyway; second, because we split a list into two
disjoint lists of packages and other entities, which is difficult to
maintain and inefficient).

The better solution is a multi-tree container of records with two multi-tree
containers each: one for packages, one for other entities. TBD.