minor removal of useless _Kind suffix in literals
Individual literals of an enumeration type should not be suffixed by _Kind,
just like we have E_Procedure and not E_Procedure_Kind for entity kinds.

Also, Ins are Inputs (because they represent the "Input" string in the
aspect syntax); the same for Outputs but not for Proof_Ins (since they
represent "Proof_In" in the aspect syntax).

Finally, Null_Kind renamed to Null_Global_Id and Null_Global_Id renamed to
Empty_Global_Id, which matches the front end convention.