P405-006 Add library of SPARK lemmas
This new library can be used in client code by setting appropriately the
environment variable SPARK_LEMMAS_OBJECT_DIR defining the object directory
for the library (as it should be in a writable area). It consists
currently of 1 lemma on modular integers (32 and 64 bits) and 5 lemmas on
signed integers (32 and 64 bits), the proof of which requires 6 manual
proofs in Coq overall. Some lemmas do not require a proof because Z3 is
able to prove them when -gnato13 is set, like it is the case here, while
the corresponding property cannot be proved in client code, thus requiring
the use of the lemma.

* include/proof/Coq/spark-integer_arithmetic_lemmas/*.v
* include/proof/Coq/spark-long_integer_arithmetic_lemmas/*.v
* include/proof/Coq/spark-mod32_arithmetic_lemmas/*.v
* include/proof/Coq/spark-mod64_arithmetic_lemmas/*.v
Manual Coq proofs for 6 VCs.

* include/proof/sessions/spark-integer_arithmetic_lemmas/why3session.xml
* include/proof/sessions/spark-integer_arithmetic_lemmas/why3shapes
* proof/sessions/spark-long_integer_arithmetic_lemmas/why3session.xml
* include/proof/sessions/spark-long_integer_arithmetic_lemmas/why3shapes
* include/proof/sessions/spark-mod32_arithmetic_lemmas/why3session.xml
* include/proof/sessions/spark-mod32_arithmetic_lemmas/why3shapes
* include/proof/sessions/spark-mod64_arithmetic_lemmas/why3session.xml
* include/proof/sessions/spark-mod64_arithmetic_lemmas/why3shapes
Session and shape files for SPARK lemma library.

* include/spark-arithmetic_lemmas.adb
* include/spark-arithmetic_lemmas.ads
* include/spark-integer_arithmetic_lemmas.ads
* include/spark-long_integer_arithmetic_lemmas.ads
* include/spark-mod32_arithmetic_lemmas.ads
* include/spark-mod64_arithmetic_lemmas.ads
* include/spark-mod_arithmetic_lemmas.adb
* include/spark-mod_arithmetic_lemmas.ads
* include/spark.ads
* include/spark_lemmas.gpr
New library of SPARK lemmas.

* why3
Update for new Coq driver.