P405-006 Add lemma for monotonicity of division on modular
Although CVC4 manages to prove this property in some cases, in other cases
it takes much too many steps to do so. Hence the need for a lemma.
Add proof without registering it in Why3 session, until P429-031 is fixed.

* Spark__mod64_arithmetic_lemmas__lemma_onic__pragargs__cmp.v
New proof (not yet registered in session or duplicated for 32bits).

* include/spark-mod_arithmetic_lemmas.adb
* include/spark-mod_arithmetic_lemmas.ads
Add lemma.

* testsuite/gnatprove/tests/P405-006__mod_arithmetic_lemmas/test_arith.adb
Call lemma.