P318-022 use lists in GG phase 1
In the first phase of generating globals the state information data is only
recorded and never queried. This is now implemented as List.Append and not
Hashed_Sets.Include (which should really be Hashed_Sets.Insert, by the way).

The complexity of GG_Register_State was O(log N); now it is O(1).