P427-045 Redefine for of quantification for proof
For ... of quantification on a container allows to
quantify on its elements rather than directly on its
cursors, providing a seemingly higher level view of
a specification.
We used to always translate this quantification back
to cursors for proof.
Now the user can choose between three translations.
If nothing is specified, like before, the quantification
is done over cursors.
If the user provides a Contains primitive, the
quantification is done directly over elements using the
Contains function.
If the user provides a Model primitive, the quantification
is translated as if it was done on the result of this
Model function.

* gnat2why/why/gnat2why-expr.adb
(Parse_Iteration_Scheme_For_Iterable): Go through Model
functions to find the container on which the iteration
should be done.
(Make_Constraint_For_Iterable): If the container has a
Contains function, then use it to construct the
constraint.
(Transform_Quantified_Expression): Take into account the
pragmas Annotate (GNATprove, Iterable_For_Proof, ...)
when translating a for ... of quantification.