minor fix of inefficient claim processing
Call to Contains followed by Insert requires two searches in the container.
Fixed as a single call to Insert and a check if the new element was indeed
inserted.

It could happen that a first token of Needed_Claims was assigned to
Cur_Token, then deleted and then included back into Needed_Claims. This
resulted in an extra (completely useless) loop iteration. Fixed by moving
Delete at the end of loop.

Tokens are now inserted (and not included) to Unverified_Claims set, which
ensures that indeed there are no useless loop iterations.

* common/assumptions-search.adb
(Claim_Depends_On): fix of inefficient set operation and potential useless
 loop iteration.