P510-003 test - enable vcg on flow-only tests
Previously many tests were run with just --mode=flow. We want to also
generate VCs for these (but not prove anything, since there should be no
interesting proof properties). This has uncovered a few crashes in proof,
so is a useful activity. For tests with do_flow we now run them in
--benchmark mode (so nothing is actually proven, but its fast).

Disabled tests that will be addressed by new tickets:
   P511-006 - proof - crash on simple extended returns
   P511-008 - proof - various crashes on very bad aliasing
   P511-009 - proof - support attribute 'alignment