P318-022 scan for blocking ops only if needed
All subprograms referenced in the current compilation units were scanned for
potentially blocking statements. Now we scan only subprograms defined in the
current compilation unit.

Also, the nonblocking subprograms are kept in a list (with unique elements),
where the complexity of the Append is O(1), which is better than O(log N)
for Insert.