P427-032 inline function postconditions for proof
When a pragma Annotate (GNATprove, Inline_For_Proof, ...)
is provided for a function entity with a post of the
form F'Result = <Expr> then translate F as a
function definition with label "inline" in Why3 so
that its definition is inlined for proof.

* gnat2why/why/gnat2why-annotate.adb
(Check_Inline_Annotation): Check validity of a pragma
Annotate Inline_For_Proof and store it.
(Retrieve_Inline_Annotation): Returns the Node_Id for
the value of an inlined function.

* gnat2why/why/gnat2why-subprograms.adb
(Compute_Inlined_Expr): Computes a Why3 expression for
a function's definition. It is non-empty only if the
function is non-recursive and has a pragma
Inline_For_Proof.
(Translate_Subprogram_Spec): Call Compute_Inlined_Expr
to construct the function's definition.
(Generate_Axiom_For_Post): Do not generate an axiom if
the function is inlined.