P421-005 flow - support pragma unused/unmodified
Added support for pragma unmodified and pragma unused (before only pragma
unreferenced was actually used). These three pragmas are treated in the
same way and a set is created containing the variables mentioned in either
of them. Warnings on variables being not used/unmodified/unreferenced but
mentioned in the pragmas are therefore silenced.