minor removal of explicit new/free in SCC code
Code for strongly connected components (SCC) used explicit memory allocation
and deallocation, which is known to prone to errors. Now it uses standard
containers.

Note: in one of the previous (minor) commits to this code I put wrong
justification about the use of arrays as big as the compiler's node array
(7a23c0b7). This was wrong: the arrays were only as big as needed to keep
the "nodes" (sic!) of the call graph.

Yet, after the rewrite with containers the code seems easier to understand,
because there are no "placeholder arrays" declared, i.e. uninitialized
memory that is only filled later. After this commit there are no such arrays
anymore.