minor move to avoid copying container's data
Use Move to not create an (implicit) copy (such a copy was not done in the
original code). Reserve the capacity exactly as it was done in the original
code with variable-size array. Now the code uses containers and not manually
managed pointers to arrays, yet it should be no less efficient.

Also, refine type that represents positive indexes.