P307-006 tool - detect missing clean-up
We now embed the version of the tools (e.g. "17.0 (20160101)" or similar)
into the ali file preceeding the globals section. We check if
a) this is present
b) if present, is it equal to the current version
This allows us to detect cases where a user might use files generated
with a previous version of SPARK or a previous wavefront.