P427-045 assume dynamic invariant of qunatified vars
In translation of quantification over types with
iterable, we were missing the dynamic invariant of the
quantified variable. This is not unsound, but, when the
cursor type has a dynamic invariant, it looses the
possibility to unfold the definition of has_element.

* gnat2why/why/gnat2why-expr.adb
(Translate_Quantified_Expression): Add the dynamic
invariant of the quantified variable to the guard of the
quantified formula.