minor removal of tests on Null_Part
When testing for Null_Flow_Scope it is enough to test the entity part, since
the invariant of type Flow_Scope implies that if the entity part is empty
then the section part is Null_Part. This is now done with Present/No
routines.

Also, explicit accesses to Section of a Null_Flow_Scope indicate flaws in
the control flow. These are guarded with calls to Present.

Also, one "and" changed to "and then", which fixes both the style and the
logic.

Note: the real trouble here was that Flow_Scope should be a discriminated
record, but discriminating on Entity_Id would result in a lot of extra code.
Perhaps leaving the Section component uninitialized by default is a better
choice (which is how the problems fixed in this commit were detected).