P518-003 remove Has_Access_Type function
Has_Access_Type goes to the SPARK representative
type to check wether it is an access type. This
cannot happen in valid SPARK code as access types
are not in SPARK.

* gnat2why/spark/spark_util.ads
(Has_Access_Type): Removed.

* gnat2why/spark/spark_definition.adb
(Mark): Use Is_Access_Type to determine if a
selected component is an implicit dereference.