P405-006 Generate correct session files
Through command-line calls to gnatprove, generate correct session files
for the SPARK lemma library, with appropriate references to generates
Coq files when manual proof is required. These are not yet marked as
proved in the session file, due to an error when calling Coq in this
setup on the Builtin generated file.

* Makefile
Add clean entry.

* generate_sessions.sh
New script to generate the starting session files.

* proof/sessions/spark-integer_arithmetic_lemmas/why3session.xml
* proof/sessions/spark-integer_arithmetic_lemmas/why3shapes
* proof/sessions/spark-long_integer_arithmetic_lemmas/why3session.xml
* proof/sessions/spark-long_integer_arithmetic_lemmas/why3shapes
* proof/sessions/spark-mod32_arithmetic_lemmas/why3session.xml
* proof/sessions/spark-mod32_arithmetic_lemmas/why3shapes
* proof/sessions/spark-mod64_arithmetic_lemmas/why3session.xml
* proof/sessions/spark-mod64_arithmetic_lemmas/why3shapes
New session and shape files.

* spark-arithmetic_lemmas.ads
* spark-mod_arithmetic_lemmas.ads
Mark lines that require manual proof.